Implement in SidebarControl.xaml.cs
